+++ pylinkage/geometry/__init__.py
@@ -2,6 +2,17 @@
 Basic geometry package.
 """
 
+__all__ = [
+    "circle_intersect",
+    "circle_line_from_points_intersection",
+    "circle_line_intersection",
+    "cyl_to_cart",
+    "get_nearest_point",
+    "intersection",
+    "norm",
+    "sqr_dist",
+]
+
 from .core import (
     cyl_to_cart as cyl_to_cart,
 )