+++ pylinkage/joints/__init__.py
@@ -1,5 +1,7 @@
 """Definition of joints."""
 
+__all__ = ["Crank", "Fixed", "Linear", "Revolute", "Static"]
+
 from .crank import Crank as Crank
 from .fixed import Fixed as Fixed
 from .joint import Static as Static