+++ pylinkage/linkage/__init__.py
@@ -1,6 +1,9 @@
 """
 Definition and analysis of a linkage as a dynamic set of joints.
 """
+
+__all__ = ["Linkage", "Simulation", "bounding_box", "kinematic_default_test"]
+
 from .analysis import (
     bounding_box as bounding_box,
 )
@@ -8,3 +11,4 @@ from .analysis import (
     kinematic_default_test as kinematic_default_test,
 )
 from .linkage import Linkage as Linkage
+from .linkage import Simulation as Simulation